Creating a real database

+++ ASPTheDepartment/Pages/Employees/Edit.cshtml.cs
@@ -55,7 +55,9 @@ namespace ASPTheDepartment.Pages
                     if (Employee.PhotoPath != null)
                     {
                         string filePath = Path.Combine(_webHostEnvironment.WebRootPath, "images", Employee.PhotoPath);
-                        System.IO.File.Delete(filePath);
+
+                        if(Employee.PhotoPath != "images.png")
+                            System.IO.File.Delete(filePath);
                     }
 
                     Employee.PhotoPath = ProcessUploadedFile();
